Move all views to their own directory grouped in seperate files
