+++ survey/views/results.py
@@ -1,13 +1,13 @@
 from django.views.generic import TemplateView, DetailView
 from django.db.models import Avg, Q
 from django.shortcuts import redirect
-from django.core.cache import cache, caches
+from django.core.cache import caches
 from enum import Enum, auto
 from collections import OrderedDict
 import inspect
 import json
-from .models import Survey, AnimeResponse, Response, SurveyAdditionRemoval, AnimeName
-from .util import SurveyUtil, get_user_info, AnimeUtil
+from survey.models import AnimeResponse, Response, SurveyAdditionRemoval
+from survey.util import SurveyUtil, get_user_info, AnimeUtil
 
 ANIME_POPULARITY_THRESHOLD = 0.02 # Anime with a popularity lower than this won't get included in results tables by default
